+++ collector/pg_setting.go
@@ -1,4 +1,4 @@
-package main
+package collector
 
 import (
 	"database/sql"
