+++ collector/postgres_exporter_test.go
@@ -1,6 +1,6 @@
 // +build !integration
 
-package main
+package collector
 
 import (
 	. "gopkg.in/check.v1"
@@ -89,11 +89,11 @@ func (s *FunctionalSuite) TestSemanticVersionColumnDiscard(c *C) {
 // test read username and password from file
 func (s *FunctionalSuite) TestEnvironmentSettingWithSecretsFiles(c *C) {
 
-	err := os.Setenv("DATA_SOURCE_USER_FILE", "./tests/username_file")
+	err := os.Setenv("DATA_SOURCE_USER_FILE", "../tests/username_file")
 	c.Assert(err, IsNil)
 	defer UnsetEnvironment(c, "DATA_SOURCE_USER_FILE")
 
-	err = os.Setenv("DATA_SOURCE_PASS_FILE", "./tests/userpass_file")
+	err = os.Setenv("DATA_SOURCE_PASS_FILE", "../tests/userpass_file")
 	c.Assert(err, IsNil)
 	defer UnsetEnvironment(c, "DATA_SOURCE_PASS_FILE")
 
@@ -103,7 +103,7 @@ func (s *FunctionalSuite) TestEnvironmentSettingWithSecretsFiles(c *C) {
 
 	var expected = "postgresql://custom_username:custom_password@localhost:5432/?sslmode=disable"
 
-	dsn := getDataSource()
+	dsn := GetDataSource()
 	if dsn != expected {
 		c.Errorf("Expected Username to be read from file. Found=%v, expected=%v", dsn, expected)
 	}
@@ -117,7 +117,7 @@ func (s *FunctionalSuite) TestEnvironmentSettingWithDns(c *C) {
 	c.Assert(err, IsNil)
 	defer UnsetEnvironment(c, "DATA_SOURCE_NAME")
 
-	dsn := getDataSource()
+	dsn := GetDataSource()
 	if dsn != envDsn {
 		c.Errorf("Expected Username to be read from file. Found=%v, expected=%v", dsn, envDsn)
 	}
@@ -131,7 +131,7 @@ func (s *FunctionalSuite) TestEnvironmentSettingWithDnsAndSecrets(c *C) {
 	c.Assert(err, IsNil)
 	defer UnsetEnvironment(c, "DATA_SOURCE_NAME")
 
-	err = os.Setenv("DATA_SOURCE_USER_FILE", "./tests/username_file")
+	err = os.Setenv("DATA_SOURCE_USER_FILE", "../tests/username_file")
 	c.Assert(err, IsNil)
 	defer UnsetEnvironment(c, "DATA_SOURCE_USER_FILE")
 
@@ -139,7 +139,7 @@ func (s *FunctionalSuite) TestEnvironmentSettingWithDnsAndSecrets(c *C) {
 	c.Assert(err, IsNil)
 	defer UnsetEnvironment(c, "DATA_SOURCE_PASS")
 
-	dsn := getDataSource()
+	dsn := GetDataSource()
 	if dsn != envDsn {
 		c.Errorf("Expected Username to be read from file. Found=%v, expected=%v", dsn, envDsn)
 	}
